+++ renderer/src/hittable.rs
@@ -1,11 +1,11 @@
 use super::{HitRecord, Ray};
 
 pub trait Hittable {
-    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord>;
+    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord>;
 }
 
 //impl<'a> Hittable for Vec<&'a dyn Hittable> {
-//    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
+//    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
 //        self.iter().fold(None, |hr, object| {
 //            if let Some(prev_hit) = &hr {
 //                match object.hit(ray, t_min, prev_hit.t) {
@@ -20,7 +20,7 @@ pub trait Hittable {
 //}
 
 impl Hittable for &[Box<dyn Hittable>] {
-    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
+    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
         self.iter().fold(None, |hr, object| {
             if let Some(prev_hit) = &hr {
                 match object.hit(ray, t_min, prev_hit.t) {
