+++ renderer/src/ray.rs
@@ -5,7 +5,7 @@ pub struct Ray {
 }
 
 impl Ray {
-    pub fn at(&self, t: f32) -> Vec3 {
+    pub fn at(&self, t: f64) -> Vec3 {
         self.origin + self.dir * t
     }
 }
